Updates ansible installer to work in container
The ansible roles were updated to use the ansible environment variable for HOME
instead of ~. The vars/main.yml file was not updated with this change and is
updated for consistency.

The kubernetes_in_cluster inventory variable is exposed allowing the ansible
installer to run from inside of a kube cluster. If this isn't set the ansible
installer requires settings to access the kubernentes or openshift clusters.
Since the container will be running from inside of the cluster with
cluster-admin permissions it doesn't need these settings

In ansible the remote_tmp directory defaults to $HOME. This directory is not
writable in a container so updating the value to the writable /tmp dir.