Adds ansible roles for pgo-installer
These ansible roles will setup, run, and cleanup the necessary resources for the
install job depending on the tags that are used and the settings in the
inventory file. The roles will run the preflight checks on the inventory file
before the job is submitted. This will help provide the use more immediate
feedback if there is an error in the inventory file. Preflight checks are also
added for specific container install checks that need to be done.

This commit also adds documentation for both the pgo-installer image and the
Ansible roles for the pgo-installer.

Updates ansible roles for postgres-operator to fix bug. Task was referencing
backrest_aws_s3_key_secret instead of backrest_aws_s3_secret.