WIP - attempt to have template handle deep yaml
The working solution is a function that ranges over the entire .Values object.
The range only works if .Values is flat and shows deeper yaml values as a single
value:
```
key: map[sub1: value1]
```

This commit is an attempt to recursively print the list so that each value is
printed as proper yaml. The issue that i'm running into is that the spacing
isn't being added correctly, ex:
```
key:
sub1: value1

key:
  sub1: value1
```

More research will be needed to handle non-flat files but this is not currently
blocking the original change because our Values yaml is flat.