Initial E2E test (#2808)
* Add initial Kuttl e2e tests

This change adds a quick end-to-end test that use the kuttl test
framework. The layout for kuttl tests is `testing/kuttl/e2e` with
individual tests in subdirectories of the `e2e` directory.

Run test(s) by using the `make check-kuttl` target.

The initial creates a cluster then checks the postgrescluster status and
that the initial backup completes.  

the subsequent test uses the above cluster and 2 scripts to run a portforward
and to connect to psql and run select version() to return the psql version


Co-authored-by: ValClarkson <valerie.clarkson@crunchydata.com>