Update exporter to use pass file
The exporter container now provides the ccp_monitoring password to
postgres_exporter using a password file instead of an environment
variable. With this, the password can be updated without requiring a
container restart. The path to the password file has also been added to
the exporter watcher logic meaning that the postgres_exporter process
will be restarted when either the queries directory or password file
change.

The password change test is updated to check that the
postgres_exporter pid has been updated before trying to re-connect.